refactor(gym/reporting): split tests to tests_reporting_extra.rs (#1266)
Existing tests_reporting.rs (752 LOC) was already a sibling; created
NEW tests_reporting_extra.rs for the inline-mod tests.

- src/gym/reporting.rs: 737 → 398 LOC
- New src/gym/tests_reporting_extra.rs (345 LOC)

Pure refactor: 34 tests pass.
